# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: KSHETRA API
  version: 0.1.0
  description: >
    Political intelligence platform API. Phase 1 serves typed seed data
    (Telangana fully supported) over a versioned REST surface. All mutating
    or privileged routes use header-based identity placeholders pending the
    Supabase JWT integration (Phase 2B).
  license:
    name: UNLICENSED (Proprietary)
servers:
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    monitorBearer:
      type: http
      scheme: bearer
      description: Shared secret for cron monitor webhooks (KSHETRA_MONITOR_SECRET).
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Service-to-service key (Phase-1 placeholder; Supabase JWT in Phase 2B).
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
      required: [error]
...
    ChatMessage:
      type: object
      properties:
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string, minLength: 1 }
      required: [role, content]
...
paths:
...
  /api/v1/ai/chat:
    post:
      tags: [ai]
      summary: Conversational political analysis (graceful without API key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/ChatMessage' }
                constituencyAcNo: { type: integer, minimum: 1, maximum: 119 }
              required: [messages]
      responses:
        '200': { description: AI response }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/ai/analyze/constituency/{acNo}:
    get:
      tags: [ai]
      summary: Quick AI analysis for a constituency
      parameters:
        - { name: acNo, in: path, required: true, schema: { type: integer, minimum: 1, maximum: 119 } }
      responses:
        '200': { description: Analysis }
        '400': { description: Invalid AC number }
...
  /api/v1/ai/smart-search:
    post:
      tags: [ai]
      summary: Natural-language constituency search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string, minLength: 3 }
              required: [query]
      responses:
        '200': { description: Ranked results }
        '400': { description: Validation error }
  /api/v1/ai/summarize-issues:
    post:
      tags: [ai]
      summary: Summarize civic issues for a constituency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                constituencyName: { type: string, minLength: 1 }
                issues: { type: array, minItems: 1, items: { type: string } }
              required: [constituencyName, issues]
      responses:
        '200': { description: Summary }
        '400': { description: Validation error }
...